refactor: Rename DocDD to DDD and update URLs
This commit renames all instances of "DocDD" to "DDD" and "Document Driven Design" to "Dialectic Driven Design" throughout the codebase.

It also updates all URLs pointing to the old GitHub organization `github.com/selberhad` to the new organization `github.com/dialecticianai`.

This change was done using a perl script to automate the process.